extract components from Tooltip
